+++ website/src/layouts/schedule/index.js
@@ -1,0 +1,14 @@
+import DashboardLayout from "../../examples/LayoutContainers/DashboardLayout";
+import DashboardNavbar from "../../examples/Navbars/DashboardNavbar";
+
+function Schedulers() {
+  return (
+    <DashboardLayout>
+      <DashboardNavbar>
+        <div>Schedulers</div>
+      </DashboardNavbar>
+    </DashboardLayout>
+  );
+}
+
+export default Schedulers;